fix: solve some code tidy issues

+++ src/ui/dialog/import_export/KeyServerImportDialog.cpp
@@ -179,7 +179,8 @@ QComboBox* KeyServerImportDialog::create_comboBox() {
       comboBox->addItem(key_server_str.c_str());
     }
 
-    int default_key_server_index = key_server_json.Check("default_server", 0);
+    size_t default_key_server_index =
+        key_server_json.Check("default_server", 0);
     if (default_key_server_index >= key_server_list.size()) {
       throw std::runtime_error("default_server index out of range");
     }
@@ -450,7 +451,8 @@ void KeyServerImportDialog::SlotImport(const KeyIdArgsListPtr& keys) {
       const auto key_server_list =
           key_server_json.Check("server_list", nlohmann::json::array());
 
-      int default_key_server_index = key_server_json.Check("default_server", 0);
+      size_t default_key_server_index =
+          key_server_json.Check("default_server", 0);
       if (default_key_server_index >= key_server_list.size()) {
         throw std::runtime_error("default_server index out of range");
       }
