+++ src/ui/thread/KeyServerImportTask.cpp
@@ -61,7 +61,7 @@ void GpgFrontend::UI::KeyServerImportTask::dealing_reply_from_server() {
   }
   emit SignalKeyServerImportResult(network_reply, buffer);
 
-  if (result_count_++ == keyids_.size() - 1) {
+  if (static_cast<size_t>(result_count_++) == keyids_.size() - 1) {
     emit SignalTaskShouldEnd(0);
   }
 }
